Added pop-ups that allows users to know when their information has been added

+++ Pages/HR/CreateAJobForm.razor.cs
@@ -55,6 +55,10 @@ namespace XebecPortal.UI.Pages.HR
         private int savedJobId;
         private int existJobId;
 
+        private bool addedNewDep;
+        private bool validUpload;
+        private bool allowedToRedirect = false;
+
         protected override async Task OnInitializedAsync()
         {            
             token = await localStorage.GetItemAsync<string>("jwt_token");
@@ -122,7 +126,16 @@ namespace XebecPortal.UI.Pages.HR
         private async Task removeDepartment(Department value)
         {
             Console.WriteLine("Value received: " + value.Id);
-            await HttpClient.DeleteJsonAsync($"https://xebecapi.azurewebsites.net/api/Department/{value.Id}", new AuthenticationHeaderValue("Bearer", token));
+            Departments.Remove(value);
+            if (await jsRuntime.InvokeAsync<bool>("confirm", "Are You Certain You Want To Remove This Department?"))
+            {
+                var removeDep = await HttpClient.DeleteJsonAsync($"https://xebecapi.azurewebsites.net/api/Department/{value.Id}", new AuthenticationHeaderValue("Bearer", token));
+                if (removeDep.IsSuccessStatusCode)
+                {
+                    await jsRuntime.InvokeAsync<object>("alert", "Department has successfully been removed!");                   
+                }
+            }
+            
             await OnInitializedAsync();
         }
 
@@ -130,9 +143,24 @@ namespace XebecPortal.UI.Pages.HR
         {
             foreach (var item in NewDepartments)
             {
-                await HttpClient.PostJsonAsync($"https://xebecapi.azurewebsites.net/api/Department", item, new AuthenticationHeaderValue("Bearer", token));
+                var newDepAdded = await HttpClient.PostJsonAsync($"https://xebecapi.azurewebsites.net/api/Department", item, new AuthenticationHeaderValue("Bearer", token));
+                if (newDepAdded.IsSuccessStatusCode)
+                {
+                    addedNewDep = true;
+                }
+                else
+                {
+                    addedNewDep = false;
+                }
+               
+            }
+
+            if (addedNewDep)
+            {
+                await jsRuntime.InvokeAsync<object>("alert", "Newly added departments has been saved!");
             }
             NewDepartments.Clear();
+            
             await createDep(false);
         }
         private void addCompany(string value)
@@ -220,21 +248,39 @@ namespace XebecPortal.UI.Pages.HR
                 if (createJobExists)
                 {
                     Console.WriteLine("Job exists already with ID of: " + existJobId);
-                    await HttpClient.PutJsonAsync($"https://xebecapi.azurewebsites.net/api/Job/{existJobId}", item, new AuthenticationHeaderValue("Bearer", token));
+                    //var validNewUpload = await HttpClient.PutJsonAsync($"https://xebecapi.azurewebsites.net/api/Job/{existJobId}", item, new AuthenticationHeaderValue("Bearer", token));
+                    //if (validNewUpload.IsSuccessStatusCode)
+                    //{
+                    //    validUpload = true;
+                    //}
+                    //else
+                    //{
+                    //    validUpload = false;
+                    //}
                     await OnInitializedAsync();
                 }
                 else
                 {
                     Console.WriteLine("Job Doesnt exist yet");
-                    await HttpClient.PostJsonAsync($"https://xebecapi.azurewebsites.net/api/Job", item, new AuthenticationHeaderValue("Bearer", token));
+                    //var validExistingUpload = await HttpClient.PostJsonAsync($"https://xebecapi.azurewebsites.net/api/Job", item, new AuthenticationHeaderValue("Bearer", token));
+                    //if (validExistingUpload.IsSuccessStatusCode)
+                    //{
+                    //    validUpload = true;
+                    //}
+                    //else
+                    //{
+                    //    validUpload = false;
+                    //}
                     await OnInitializedAsync();
                 }
 
-                // create a check statement here
-                //await HttpClient.PostJsonAsync($"https://xebecapi.azurewebsites.net/api/Job", item, new AuthenticationHeaderValue("Bearer", token));
-
-                //await httpClient.PutJsonAsync($"https://xebecapi.azurewebsites.net/api/PersonalInformation/{item.Id}", item, new AuthenticationHeaderValue("Bearer", token));
             }
+
+            if (validUpload)
+            {
+                await jsRuntime.InvokeAsync<object>("alert", "The current state of the job creation process has been saved!");
+            }
+            allowedToRedirect = true;
             jobList.Clear();
             jobType.Clear();
             checkJobList.Clear();
